+++ lending/RELEASES.md
@@ -29,3 +29,13 @@ Based on:
 - [python v3.1.0] lending
 ### Releases
 - [PyPI v3.1.0] https://pypi.org/project/codat-lending/3.1.0 - lending
+
+## 2023-09-18 21:44:51
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.87.0 (2.116.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.1.0] lending
+### Releases
+- [PyPI v4.1.0] https://pypi.org/project/codat-lending/4.1.0 - lending